Separate plugin and processor
